requirements.txt
Django==1.11.12
dj_database_url==0.5.0
whitenoise==3.3.1
djangorestframework==3.8.2
gunicorn==19.8.1
